@@ -7,9 +7,9 @@ import todayquest.common.MessageUtil;
 import todayquest.reward.entity.Reward;
 import todayquest.user.entity.UserInfo;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
-import static javax.persistence.FetchType.LAZY;
+import static jakarta.persistence.FetchType.LAZY;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
