@@ -9,8 +9,8 @@ import lombok.RequiredArgsConstructor;
 import todayquest.item.entity.QItemLog;
 import todayquest.quest.entity.QQuestLog;
 
-import javax.annotation.PostConstruct;
-import javax.persistence.EntityManager;
+import jakarta.annotation.PostConstruct;
+import jakarta.persistence.EntityManager;
 import java.util.Map;
 import java.util.stream.Collectors;
 
