@@ -7,9 +7,9 @@ import todayquest.reward.entity.Reward;
 import todayquest.reward.entity.RewardGrade;
 import todayquest.user.entity.UserInfo;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 
 
 @Data
