@@ -2,8 +2,8 @@ package todayquest.user.dto;
 
 import lombok.Data;
 
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
 
 @Data
 public class UserRequestDto {
